+++ src/fieldlayoutelements/mailinglists/MailingListContactFieldLayoutElement.php
@@ -8,6 +8,7 @@ namespace putyourlightson\campaign\fieldlayoutelements\mailinglists;
 use Craft;
 use craft\base\ElementInterface;
 use craft\fieldlayoutelements\BaseNativeField;
+use craft\helpers\UrlHelper;
 use putyourlightson\campaign\assets\ContactEditAsset;
 use putyourlightson\campaign\elements\ContactElement;
 use putyourlightson\campaign\elements\MailingListElement;
@@ -45,6 +46,16 @@ class MailingListContactFieldLayoutElement extends BaseNativeField
         $contacts = $contactsQuery->all();
         $total = $contactsQuery->count();
 
+        $viewAllUrl = UrlHelper::cpUrl('campaign/contacts/view', [
+            'site' => Craft::$app->getSites()->getCurrentSite()->handle,
+            'source' => 'mailingList:' . $element->uid,
+        ]);
+
+        // TODO: remove in 3.0.0, when element index URLs include the source, added in Craft 4.3.0.
+        if (version_compare(Craft::$app->getVersion(), '4.3.0', '<')) {
+            $viewAllUrl = UrlHelper::cpUrl('campaign/contacts/view/' . $element->id);
+        }
+
         return Craft::$app->getView()->renderTemplate(
             'campaign/mailinglists/_includes/contacts',
             [
@@ -52,6 +63,7 @@ class MailingListContactFieldLayoutElement extends BaseNativeField
                 'contacts' => $contacts,
                 'total' => $total,
                 'limit' => $limit,
+                'viewAllUrl' => $viewAllUrl,
             ],
         );
     }